# tests/snapshots/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ConfirmDecomposeBootstrapWorkflowRequest": {
        "properties": {
          "approved": {
            "default": true,
            "title": "Approved",
            "type": "boolean"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "confirmed_by": {
            "minLength": 1,
            "title": "Confirmed By",
            "type": "string"
          },
          "expected_modules": {
            "items": {
              "type": "string"
            },
            "title": "Expected Modules",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "confirmed_by"
        ],
        "title": "ConfirmDecomposeBootstrapWorkflowRequest",
        "type": "object"
      },
      "ConfirmDecomposeBootstrapWorkflowResponse": {
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "confirmation_status": {
            "title": "Confirmation Status",
            "type": "string"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "confirmed_by": {
            "title": "Confirmed By",
            "type": "string"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "workitems": {
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            },
            "title": "Workitems",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "approved",
          "confirmation_status",
          "confirmed_by"
        ],
        "title": "ConfirmDecomposeBootstrapWorkflowResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapAdvanceLoopRequest": {
        "properties": {
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "execute_max_loops": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Execute Max Loops",
            "type": "integer"
          },
          "expected_modules": {
            "items": {
              "type": "string"
            },
            "title": "Expected Modules",
            "type": "array"
          },
          "force_refresh_preview": {
            "default": false,
            "title": "Force Refresh Preview",
            "type": "boolean"
          },
          "max_steps": {
            "default": 8,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "stop_when_bootstrap_finished": {
            "default": true,
            "title": "Stop When Bootstrap Finished",
            "type": "boolean"
          }
        },
        "title": "DecomposeBootstrapAdvanceLoopRequest",
        "type": "object"
      },
      "DecomposeBootstrapAdvanceLoopResponse": {
        "properties": {
          "action_status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Action Status Counts",
            "type": "object"
          },
          "action_taken_sequence": {
            "items": {
              "type": "string"
            },
            "title": "Action Taken Sequence",
            "type": "array"
          },
          "final_status": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          },
          "halted_reason": {
            "title": "Halted Reason",
            "type": "string"
          },
          "last_action_taken": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action Taken"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/DecomposeBootstrapAdvanceResponse"
            },
            "title": "Steps",
            "type": "array"
          },
          "steps_executed": {
            "default": 0,
            "title": "Steps Executed",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "halted_reason",
          "final_status"
        ],
        "title": "DecomposeBootstrapAdvanceLoopResponse",
        "type": "object"
      },
      "DecomposeBootstrapAdvanceRequest": {
        "properties": {
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "expected_modules": {
            "items": {
              "type": "string"
            },
            "title": "Expected Modules",
            "type": "array"
          },
          "force_refresh_preview": {
            "default": false,
            "title": "Force Refresh Preview",
            "type": "boolean"
          },
          "max_loops": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Loops",
            "type": "integer"
          }
        },
        "title": "DecomposeBootstrapAdvanceRequest",
        "type": "object"
      },
      "DecomposeBootstrapAdvanceResponse": {
        "properties": {
          "action_status": {
            "title": "Action Status",
            "type": "string"
          },
          "action_taken": {
            "title": "Action Taken",
            "type": "string"
          },
          "confirmation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfirmDecomposeBootstrapWorkflowResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "execute": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecuteWorkflowRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "preview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecomposeBootstrapPreviewResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status_after": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          },
          "status_before": {
            "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
          }
        },
        "required": [
          "run_id",
          "action_taken",
          "action_status",
          "status_before",
          "status_after"
        ],
        "title": "DecomposeBootstrapAdvanceResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapPreviewResponse": {
        "properties": {
          "cache_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Fingerprint"
          },
          "cache_hit": {
            "default": false,
            "title": "Cache Hit",
            "type": "boolean"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "parallel_groups": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Parallel Groups",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "task_count": {
            "default": 0,
            "title": "Task Count",
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/DecomposeBootstrapPreviewTask"
            },
            "title": "Tasks",
            "type": "array"
          },
          "terminal_task_keys": {
            "items": {
              "type": "string"
            },
            "title": "Terminal Task Keys",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "source"
        ],
        "title": "DecomposeBootstrapPreviewResponse",
        "type": "object"
      },
...
      "DecomposeBootstrapWorkflowRequest": {
        "properties": {
          "max_modules": {
            "default": 6,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Modules",
            "type": "integer"
          },
          "module_hints": {
            "items": {
              "type": "string"
            },
            "title": "Module Hints",
            "type": "array"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "requirements": {
            "minLength": 1,
            "title": "Requirements",
            "type": "string"
          }
        },
        "required": [
          "requirements"
        ],
        "title": "DecomposeBootstrapWorkflowRequest",
        "type": "object"
      },
      "DecomposeBootstrapWorkflowResponse": {
        "properties": {
          "chief_agent": {
            "title": "Chief Agent",
            "type": "string"
          },
          "chief_metadata": {
            "additionalProperties": true,
            "title": "Chief Metadata",
            "type": "object"
          },
          "chief_summary": {
            "title": "Chief Summary",
            "type": "string"
          },
          "chief_trace_id": {
            "title": "Chief Trace Id",
            "type": "string"
          },
          "confirmation_required": {
            "default": false,
            "title": "Confirmation Required",
            "type": "boolean"
          },
          "confirmation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Status"
          },
          "confirmation_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmation Token"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "title": "Modules",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "workitems": {
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            },
            "title": "Workitems",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "chief_summary",
          "chief_agent",
          "chief_trace_id"
        ],
        "title": "DecomposeBootstrapWorkflowResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WhereCode Control Center",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v3/workflows/runs/{run_id}/decompose-bootstrap": {
      "post": {
        "operationId": "decompose_bootstrap_workflow_run_v3_workflows_runs__run_id__decompose_bootstrap_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecomposeBootstrapWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapWorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decompose Bootstrap Workflow Run"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/advance": {
      "post": {
        "operationId": "advance_decompose_bootstrap_run_v3_workflows_runs__run_id__decompose_bootstrap_advance_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecomposeBootstrapAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapAdvanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Decompose Bootstrap Run"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/advance-loop": {
      "post": {
        "operationId": "advance_decompose_bootstrap_run_loop_v3_workflows_runs__run_id__decompose_bootstrap_advance_loop_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecomposeBootstrapAdvanceLoopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapAdvanceLoopResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Advance Decompose Bootstrap Run Loop"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/confirm": {
      "post": {
        "operationId": "confirm_decompose_bootstrap_workflow_run_v3_workflows_runs__run_id__decompose_bootstrap_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmDecomposeBootstrapWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmDecomposeBootstrapWorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Decompose Bootstrap Workflow Run"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/pending": {
      "get": {
        "operationId": "get_decompose_bootstrap_pending_v3_workflows_runs__run_id__decompose_bootstrap_pending_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapPendingWorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decompose Bootstrap Pending"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/preview": {
      "get": {
        "operationId": "get_decompose_bootstrap_preview_v3_workflows_runs__run_id__decompose_bootstrap_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decompose Bootstrap Preview"
      }
    },
    "/v3/workflows/runs/{run_id}/decompose-bootstrap/status": {
      "get": {
        "operationId": "get_decompose_bootstrap_aggregate_status_v3_workflows_runs__run_id__decompose_bootstrap_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecomposeBootstrapAggregateStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Decompose Bootstrap Aggregate Status"
      }
    },
...
    "/v3/workflows/runs/{run_id}/orchestrate": {
      "post": {
        "operationId": "orchestrate_workflow_run_v3_workflows_runs__run_id__orchestrate_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRunOrchestrateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunOrchestrateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orchestrate Workflow Run"
      }
    },
    "/v3/workflows/runs/{run_id}/orchestrate/latest": {
      "get": {
        "operationId": "get_latest_orchestrate_telemetry_v3_workflows_runs__run_id__orchestrate_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunOrchestrateLatestTelemetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Orchestrate Telemetry"
      }
    },
    "/v3/workflows/runs/{run_id}/orchestrate/recover": {
      "post": {
        "operationId": "execute_orchestrate_recovery_action_v3_workflows_runs__run_id__orchestrate_recover_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRunOrchestrateRecoveryExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunOrchestrateRecoveryExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Orchestrate Recovery Action"
      }
    },
...
    "/v3/workflows/runs/{run_id}/routing-decisions": {
      "get": {
        "operationId": "get_workflow_run_routing_decisions_v3_workflows_runs__run_id__routing_decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunRoutingDecisionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Run Routing Decisions"
      }
    },
...
  }
}